refact: #67 separate common message styles into @styles/messages.js
